supposedly fixed with Casper v1.3.1
Ported this SC-API commit[0] to Snapy which supposedly fixes the login
issues. I personally however am still getting the following error:

"Oh no! Your login temporarily failed, so please try again later. If
your login continues to fail, please visit
https://support.snapchat.com/a/failed-login :)"

Some people on the Casper Facebook thread[1] are saying they are still
experiencing this error as well, so I'm not sure if I made a mistake
porting the SC-API commit or if the fault is on Casper's side.

[0] https://github.com/mgp25/SC-API/commit/b6576561ca2ae40ac5b72ac33c5339de65815780
[1] https://www.facebook.com/casper4snapchat/posts/1614555802144008